added sweet alert

+++ src/app/activite-dashboard/activite-dashboard.component.ts
@@ -4,7 +4,8 @@ import {MatPaginator} from "@angular/material/paginator";
 import {MatDialog} from "@angular/material/dialog";
 import {Activity} from "../_interfaces/activity";
 import {ActivityService} from "../_services/activity.service";
-import {ActivityDialogComponent} from "../_dialogs/activity-dialog/activity-dialog.component";
+import { ActivityDialogComponent } from "../_dialogs/activity-dialog/activity-dialog.component";
+import Swal from 'sweetalert2/dist/sweetalert2.all.js';
 
 @Component({
   selector: 'app-activite-dashboard',
@@ -38,11 +39,47 @@ export class ActiviteDashboardComponent implements OnInit {
     });
   }
 
+  //removeActivity(id: number) {
+  //  console.log(id);
+  //  this.ActivityService.DeleteActivity(id).subscribe(data => {
+  //    this.ngOnInit();
+  //  });
+  //}
+
   removeActivity(id: number) {
-    console.log(id);
-    this.ActivityService.DeleteActivity(id).subscribe(data => {
-      this.ngOnInit();
-    });
+    Swal.fire({
+      title: 'Are you sure want to remove this record?',
+      text: 'You will not be able to recover it afterwards!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, keep it'
+    }).then((result) => {
+      if (result.value) {
+        this.ActivityService.DeleteActivity(id).subscribe(data => {
+          this.dataSource.data = data;
+        });
+        Swal.fire({
+          toast: true,
+          position: 'top-end',
+          icon: 'success',
+          title: 'Record Deleted',
+          showConfirmButton: false,
+          timer: 1800
+        })
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire({
+          toast: true,
+          position: 'top-end',
+          icon: 'error',
+          title: 'Cancelled Deletion',
+          text: 'Your record will remain!',
+          showConfirmButton: false,
+          timer: 1800
+        })
+      }
+    })
   }
 
   OpenDialogToUpdate(element) {
